# api-docs/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API to manage projects",
        "title": "Tezos Link API",
        "contact": {
            "name": "API Support",
            "email": "email@ded.fr"
        },
        "license": {},
        "version": "v1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/health": {
            "get": {
                "summary": "get application health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Health"
                        }
                    }
                }
            }
        },
        "/projects": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "Create a Project",
                "parameters": [
                    {
                        "description": "New Project",
                        "name": "new-project",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/inputs.NewProject"
                        }
                    }
                ],
                "responses": {
                    "201": {},
                    "400": {}
                }
            }
        },
        "/projects/{uuid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Get a Project with the associated metrics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/outputs.ProjectOutputWithMetrics"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "inputs.NewProject": {
            "type": "object",
            "properties": {
                "title": {
                    "type": "string"
                }
            }
        },
        "model.Health": {
            "type": "object",
            "properties": {
                "connectedToDb": {
                    "type": "boolean"
                }
            }
        },
        "outputs.MetricsOutput": {
            "type": "object",
            "properties": {
                "requestsCount": {
                    "type": "integer"
                }
            }
        },
        "outputs.ProjectOutputWithMetrics": {
            "type": "object",
            "properties": {
                "metrics": {
                    "type": "object",
                    "$ref": "#/definitions/outputs.MetricsOutput"
                },
                "title": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    }
}